requirements-macropad.txt
# HX-K12 macro pad (PC Windows)
hidapi
keyboard
pyyaml
